Update and rename library.java to Library.java
